LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/precommitHandler-traversal-window-stop-before-commit-expected.txt


FAIL  precommitHandler traverse with window.stop() before commit assert_equals: expected "#2" but got "#1"
